+++ tsdb/engine/tsm1/engine_test.go
@@ -783,6 +783,7 @@ type mockPlanner struct{}
 
 func (m *mockPlanner) Plan(lastWrite time.Time) []tsm1.CompactionGroup { return nil }
 func (m *mockPlanner) PlanLevel(level int) []tsm1.CompactionGroup      { return nil }
+func (m *mockPlanner) PlanOptimize() []tsm1.CompactionGroup            { return nil }
 
 // ParseTags returns an instance of Tags for a comma-delimited list of key/values.
 func ParseTags(s string) influxql.Tags {